Put start_year in its own R script. Created a flag for schema_nuke that determines if all schema are dropped or just the years. Created functions to trucate & nuke the master schema.

+++ Schema_dropper.R
@@ -14,8 +14,10 @@ for(i in 1:length(years)){
   dbExecute(conn = dbconnection_local, statement = script)
 }
 
-print('formula1')
-script <- paste0('DROP SCHEMA IF EXISTS formula1;')
-dbExecute(conn = dbconnection_local, statement = script)
+if(all_schema){
+  print('formula1')
+  script <- paste0('DROP SCHEMA IF EXISTS formula1;')
+  dbExecute(conn = dbconnection_local, statement = script)
+}
 
 dbKillConnections()
